refactor(api-download): move `MatrixProfile` and specifics matrix in `matrix_profile` module
